+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix playout builder crash with improperly configured pad filler preset
 - Properly validate filler preset mode pad to require `filler pad to nearest minute` value
 - Fix bug where previously-synchronized collection tags would disappear
-  - This bug affected Jellyfin, Emby and Plex collections 
+  - This bug affected Jellyfin, Emby and Plex collections
+- Fix detection of AMF hardware acceleration on Windows 
 
 ## [0.8.3-beta] - 2023-11-22
 ### Added